// src/model/GetBreakdownMetricsRequest.ts
/**
 * @aiozstream/nodejs-client
 * Aioz Stream Service
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated.
 * Do not edit the class manually.
 */

import AttributeType from './AttributeType.js';
import MetricFilter from './MetricFilter.js';

export default class GetBreakdownMetricsRequest {
  'filterBy'?: MetricFilter;
  'from'?: number;
  'limit'?: number;
  'offset'?: number;
  'orderBy'?: string;
  'sortBy'?: string;
  'sumOthers'?: boolean;
  'to'?: number;

  static readonly discriminator?: string = undefined;

  static readonly attributeTypeMap: Array<AttributeType> = [
    {
      name: 'filterBy',
      baseName: 'filter_by',
      type: 'MetricFilter',
      format: '',
    },
    {
      name: 'from',
      baseName: 'from',
      type: 'number',
      format: '',
    },
    {
      name: 'limit',
      baseName: 'limit',
      type: 'number',
      format: '',
    },
    {
      name: 'offset',
      baseName: 'offset',
      type: 'number',
      format: '',
    },
    {
      name: 'orderBy',
      baseName: 'order_by',
      type: 'string',
      format: '',
    },
    {
      name: 'sortBy',
      baseName: 'sort_by',
      type: 'string',
      format: '',
    },
    {
      name: 'sumOthers',
      baseName: 'sum_others',
      type: 'boolean',
      format: '',
    },
    {
      name: 'to',
      baseName: 'to',
      type: 'number',
      format: '',
    },
  ];

  static getAttributeTypeMap(): Array<AttributeType> {
    return GetBreakdownMetricsRequest.attributeTypeMap;
  }
}